create files for different javascript classes
