+++ book.js
@@ -1,0 +1,7 @@
+// Book class: Represents a book
+export default class Book {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+  }
+}
