+++ index.html
@@ -44,6 +44,6 @@
         <input type="submit" value="add-book" class="btn" />
       </form>
     </section>
-    <script defer src="main.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
